Crate slog_nursery

Nursery for slog-rs

This crate is forever unstable, and contains things that at a given moment are useful but not final.

Structs

Failover
Drain that switches destination of error

Functions

failover
Failover logging to secondary drain on primary's failure