Struct slog_nursery::Failover

pub struct Failover<D1: Drain, D2: Drain> { /* fields omitted */ }

Drain that switches destination of error

Logs everything to drain D1, but in case of it reporting an error, switching to D2. If D2 returns an error too, Failover will return an error.

Methods

impl<D1: Drain, D2: Drain> Failover<D1, D2>

fn new(drain1: D1, drain2: D2) -> Self

Create Failover